nrunner: split avocado/core/nrunner.py
The "avocado.core.nrunner" was implemented as a standalone
module/executable so that it could be executed on systems/environments
that have nothing but Python.  The podman spawner is one component
that used to rely on the ability to deploy that one file to a
Python capable container.

With the recent work that switch to the deployment of Avocado using
eggs, we no longer need to keep the basic nrunner as a single
standalone script.

The approach taken here was to focus on the split, keeping as much as
the same names as possible, only introducing new levels because that
was necessary for the actual split.  Also, some of the runnes for some
test types have been moved into the existing runners directory.  In
summary:

 avocado/core/nrunner.py ---> avocado/core/nrunner/$module.py
                         \
                          `-> avocado/core/runners/$runner.py

The "python -m avocado.core.nrunner" entrypoint was kept (with the
introduction of avocado/core/nrunner/__main__.py).
