+++ avocado/core/runners/avocado_instrumented.py
@@ -4,7 +4,10 @@ import tempfile
 import time
 import traceback
 
-from avocado.core import nrunner
+from avocado.core.nrunner.app import BaseRunnerApp
+from avocado.core.nrunner.runner import (RUNNER_RUN_CHECK_INTERVAL,
+                                         RUNNER_RUN_STATUS_INTERVAL,
+                                         BaseRunner)
 from avocado.core.runners.utils import messages
 from avocado.core.test import TestID
 from avocado.core.tree import TreeNodeEnvOnly
@@ -12,7 +15,7 @@ from avocado.core.utils import loader
 from avocado.core.varianter import is_empty_variant
 
 
-class AvocadoInstrumentedTestRunner(nrunner.BaseRunner):
+class AvocadoInstrumentedTestRunner(BaseRunner):
     """
     Runner for Avocado INSTRUMENTED tests
 
@@ -116,12 +119,12 @@ class AvocadoInstrumentedTestRunner(nrunner.BaseRunner):
             timeout = float(self.DEFAULT_TIMEOUT)
             most_current_execution_state_time = None
             while True:
-                time.sleep(nrunner.RUNNER_RUN_CHECK_INTERVAL)
+                time.sleep(RUNNER_RUN_CHECK_INTERVAL)
                 now = time.monotonic()
                 if queue.empty():
                     if most_current_execution_state_time is not None:
                         next_execution_state_mark = (most_current_execution_state_time +
-                                                     nrunner.RUNNER_RUN_STATUS_INTERVAL)
+                                                     RUNNER_RUN_STATUS_INTERVAL)
                     if (most_current_execution_state_time is None or
                             now > next_execution_state_mark):
                         most_current_execution_state_time = now
@@ -145,14 +148,15 @@ class AvocadoInstrumentedTestRunner(nrunner.BaseRunner):
             yield messages.FinishedMessage.get('error', fail_reason=str(e))
 
 
-class RunnerApp(nrunner.BaseRunnerApp):
+class RunnerApp(BaseRunnerApp):
     PROG_NAME = 'avocado-runner-avocado-instrumented'
     PROG_DESCRIPTION = 'nrunner application for avocado-instrumented tests'
     RUNNABLE_KINDS_CAPABLE = ['avocado-instrumented']
 
 
 def main():
-    nrunner.main(RunnerApp)
+    app = RunnerApp(print)
+    app.run()
 
 
 if __name__ == '__main__':
