+++ avocado/core/runners/tap.py
@@ -1,11 +1,12 @@
 import io
 
-from avocado.core import nrunner
+from avocado.core.nrunner.app import BaseRunnerApp
+from avocado.core.runners.exec_test import ExecTestRunner
 from avocado.core.runners.utils.messages import FinishedMessage
 from avocado.core.tapparser import TapParser, TestResult
 
 
-class TAPRunner(nrunner.ExecTestRunner):
+class TAPRunner(ExecTestRunner):
     """Runner for standalone executables treated as TAP
 
     When creating the Runnable, use the following attributes:
@@ -64,7 +65,7 @@ class TAPRunner(nrunner.ExecTestRunner):
                                    returncode=process.returncode)
 
 
-class RunnerApp(nrunner.BaseRunnerApp):
+class RunnerApp(BaseRunnerApp):
     PROG_NAME = 'avocado-runner-tap'
     PROG_DESCRIPTION = ('nrunner application for executable tests that '
                         'produce TAP')
@@ -72,7 +73,8 @@ class RunnerApp(nrunner.BaseRunnerApp):
 
 
 def main():
-    nrunner.main(RunnerApp)
+    app = RunnerApp(print)
+    app.run()
 
 
 if __name__ == '__main__':
