+++ examples/jobs/passjob_custom.py
@@ -3,7 +3,7 @@
 import sys
 
 from avocado.core.job import Job
-from avocado.core.nrunner import Runnable
+from avocado.core.nrunner.runnable import Runnable
 from avocado.core.suite import TestSuite
 
 # Custom method (no discovery, no guess, no magic)
