+++ examples/plugins/tests/magic/avocado_magic/resolver.py
@@ -16,7 +16,7 @@
 Test resolver for magic test words
 """
 
-from avocado.core.nrunner import Runnable
+from avocado.core.nrunner.runnable import Runnable
 from avocado.core.plugin_interfaces import Discoverer, Resolver
 from avocado.core.resolver import (ReferenceResolution,
                                    ReferenceResolutionResult)
