+++ examples/plugins/tests/magic/avocado_magic/runner.py
@@ -1,8 +1,9 @@
-from avocado.core import nrunner
+from avocado.core.nrunner.app import BaseRunnerApp
+from avocado.core.nrunner.runner import BaseRunner
 from avocado.core.runners.utils.messages import FinishedMessage, StartedMessage
 
 
-class MagicRunner(nrunner.BaseRunner):
+class MagicRunner(BaseRunner):
     """Runner for magic words
 
     When creating the Runnable, use the following attributes:
@@ -30,14 +31,15 @@ class MagicRunner(nrunner.BaseRunner):
         yield FinishedMessage.get(result)
 
 
-class RunnerApp(nrunner.BaseRunnerApp):
+class RunnerApp(BaseRunnerApp):
     PROG_NAME = 'avocado-runner-magic'
     PROG_DESCRIPTION = 'nrunner application for magic tests'
     RUNNABLE_KINDS_CAPABLE = ['magic']
 
 
 def main():
-    nrunner.main(RunnerApp)
+    app = RunnerApp(print)
+    app.run()
 
 
 if __name__ == '__main__':
