+++ optional_plugins/golang/avocado_golang/golang.py
@@ -21,7 +21,7 @@ import glob
 import os
 import re
 
-from avocado.core.nrunner import Runnable
+from avocado.core.nrunner.runnable import Runnable
 from avocado.core.plugin_interfaces import Resolver
 from avocado.core.resolver import (ReferenceResolution,
                                    ReferenceResolutionResult)
