+++ selftests/unit/test_dependencies_resolver.py
@@ -1,7 +1,7 @@
 import unittest
 
 from avocado.core.dependencies.resolver import DependencyResolver
-from avocado.core.nrunner import Runnable
+from avocado.core.nrunner.runnable import Runnable
 
 
 class BasicTests(unittest.TestCase):
