+++ selftests/unit/test_nrunner.py
@@ -3,25 +3,26 @@ import sys
 import tempfile
 import unittest.mock
 
-from avocado.core import nrunner
+from avocado.core.nrunner.runnable import Runnable
+from avocado.core.nrunner.task import Task
 from avocado.core.runners import tap as runner_tap
 from selftests.utils import skipUnlessPathExists, temp_dir_prefix
 
 
-class Runnable(unittest.TestCase):
+class RunnableTest(unittest.TestCase):
 
     def test_runnable_args(self):
-        runnable = nrunner.Runnable('noop', 'uri', 'arg1', 'arg2')
+        runnable = Runnable('noop', 'uri', 'arg1', 'arg2')
         self.assertIn('arg1', runnable.args)
         self.assertIn('arg2', runnable.args)
 
     def test_runnable_kwargs(self):
-        runnable = nrunner.Runnable('noop', 'uri', key1='val1', key2='val2')
+        runnable = Runnable('noop', 'uri', key1='val1', key2='val2')
         self.assertEqual(runnable.kwargs.get('key1'), 'val1')
         self.assertEqual(runnable.kwargs.get('key2'), 'val2')
 
     def test_runnable_args_kwargs(self):
-        runnable = nrunner.Runnable('noop', 'uri', 'arg1', 'arg2',
+        runnable = Runnable('noop', 'uri', 'arg1', 'arg2',
                                     key1='val1', key2='val2')
         self.assertIn('arg1', runnable.args)
         self.assertIn('arg2', runnable.args)
@@ -29,13 +30,13 @@ class Runnable(unittest.TestCase):
         self.assertEqual(runnable.kwargs.get('key2'), 'val2')
 
     def test_runnable_tags(self):
-        runnable = nrunner.Runnable('noop', 'uri',
+        runnable = Runnable('noop', 'uri',
                                     tags={'arch': set(['x86_64', 'ppc64'])})
         self.assertIn('x86_64', runnable.tags.get('arch'))
         self.assertIn('ppc64', runnable.tags.get('arch'))
 
     def test_runnable_args_kwargs_tags(self):
-        runnable = nrunner.Runnable('noop', 'uri', 'arg1', 'arg2',
+        runnable = Runnable('noop', 'uri', 'arg1', 'arg2',
                                     tags={'arch': set(['x86_64', 'ppc64'])},
                                     non_standard_option='non_standard_value')
         self.assertIn('arg1', runnable.args)
@@ -46,16 +47,16 @@ class Runnable(unittest.TestCase):
                          'non_standard_value')
 
     def test_kind_required(self):
-        self.assertRaises(TypeError, nrunner.Runnable)
+        self.assertRaises(TypeError, Runnable)
 
     def test_kind_noop(self):
-        runnable = nrunner.Runnable('noop', None)
+        runnable = Runnable('noop', None)
         self.assertEqual(runnable.kind, 'noop')
 
     def test_recipe_noop(self):
         open_mocked = unittest.mock.mock_open(read_data='{"kind": "noop"}')
         with unittest.mock.patch("builtins.open", open_mocked):
-            runnable = nrunner.Runnable.from_recipe("fake_path")
+            runnable = Runnable.from_recipe("fake_path")
         self.assertEqual(runnable.kind, "noop")
 
     def test_recipe_exec(self):
@@ -64,7 +65,7 @@ class Runnable(unittest.TestCase):
                        '"args": ["/etc/profile"], '
                        '"kwargs": {"TERM": "vt3270"}}'))
         with unittest.mock.patch("builtins.open", open_mocked):
-            runnable = nrunner.Runnable.from_recipe("fake_path")
+            runnable = Runnable.from_recipe("fake_path")
         self.assertEqual(runnable.kind, "exec")
         self.assertEqual(runnable.uri, "/bin/sh")
         self.assertEqual(runnable.args, ("/etc/profile",))
@@ -72,25 +73,25 @@ class Runnable(unittest.TestCase):
 
     def test_identifier_args(self):
         config = {'runner.identifier_format': '{uri}-{args[0]}'}
-        runnable = nrunner.Runnable('exec-text', 'uri', 'arg1', 'arg2',
-                                    config=config)
+        runnable = Runnable('exec-text', 'uri', 'arg1', 'arg2',
+                            config=config)
         self.assertEqual(runnable.identifier, 'uri-arg1')
 
     def test_runnable_command_args(self):
-        runnable = nrunner.Runnable('noop', 'uri', 'arg1', 'arg2')
+        runnable = Runnable('noop', 'uri', 'arg1', 'arg2')
         actual_args = runnable.get_command_args()
         exp_args = ['-k', 'noop', '-u', 'uri', '-a', 'arg1', '-a', 'arg2']
         self.assertEqual(actual_args, exp_args)
 
     def test_get_dict(self):
-        runnable = nrunner.Runnable('noop', '_uri_', 'arg1', 'arg2')
+        runnable = Runnable('noop', '_uri_', 'arg1', 'arg2')
         self.assertEqual(runnable.get_dict(),
                          {'kind': 'noop', 'uri': '_uri_',
                           'args': ('arg1', 'arg2'),
                           'config': {}})
 
     def test_get_json(self):
-        runnable = nrunner.Runnable('noop', '_uri_', 'arg1', 'arg2')
+        runnable = Runnable('noop', '_uri_', 'arg1', 'arg2')
         expected = ('{"kind": "noop", '
                     '"uri": "_uri_", '
                     '"config": {}, '
@@ -98,7 +99,7 @@ class Runnable(unittest.TestCase):
         self.assertEqual(runnable.get_json(), expected)
 
     def test_runner_from_runnable_error(self):
-        runnable = nrunner.Runnable('unsupported_kind', '')
+        runnable = Runnable('unsupported_kind', '')
         try:
             runnable.pick_runner_class()
         except ValueError as e:
@@ -109,14 +110,14 @@ class RunnableFromCommandLineArgs(unittest.TestCase):
 
     def test_noop(self):
         parsed_args = {'kind': 'noop', 'uri': None}
-        runnable = nrunner.Runnable.from_args(parsed_args)
+        runnable = Runnable.from_args(parsed_args)
         self.assertEqual(runnable.kind, 'noop')
         self.assertIsNone(runnable.uri)
 
     def test_exec_args(self):
         parsed_args = {'kind': 'exec', 'uri': '/path/to/executable',
                        'arg': ['-a', '-b', '-c']}
-        runnable = nrunner.Runnable.from_args(parsed_args)
+        runnable = Runnable.from_args(parsed_args)
         self.assertEqual(runnable.kind, 'exec')
         self.assertEqual(runnable.uri, '/path/to/executable')
         self.assertEqual(runnable.args, ('-a', '-b', '-c'))
@@ -126,7 +127,7 @@ class RunnableFromCommandLineArgs(unittest.TestCase):
         parsed_args = {'kind': 'exec', 'uri': '/path/to/executable',
                        'arg': ['-a', '-b', '-c'],
                        'kwargs': [('DEBUG', '1'), ('LC_ALL', 'C')]}
-        runnable = nrunner.Runnable.from_args(parsed_args)
+        runnable = Runnable.from_args(parsed_args)
         self.assertEqual(runnable.kind, 'exec')
         self.assertEqual(runnable.uri, '/path/to/executable')
         self.assertEqual(runnable.args, ('-a', '-b', '-c'))
@@ -136,7 +137,7 @@ class RunnableFromCommandLineArgs(unittest.TestCase):
     def test_kwargs_json_empty_dict(self):
         parsed_args = {'kind': 'noop', 'uri': None,
                        'kwargs': [('empty', 'json:{}')]}
-        runnable = nrunner.Runnable.from_args(parsed_args)
+        runnable = Runnable.from_args(parsed_args)
         self.assertEqual(runnable.kind, 'noop')
         self.assertIsNone(runnable.uri)
         self.assertEqual(runnable.kwargs.get('empty'), {})
@@ -147,7 +148,7 @@ class RunnableFromCommandLineArgs(unittest.TestCase):
             'kwargs': [('tags', 'json:{"arch": ["x86_64", "ppc64"]}'),
                        ('hi', 'json:"hello"')]
         }
-        runnable = nrunner.Runnable.from_args(parsed_args)
+        runnable = Runnable.from_args(parsed_args)
         self.assertEqual(runnable.kind, 'noop')
         self.assertIsNone(runnable.uri)
         self.assertEqual(runnable.kwargs.get('hi'), 'hello')
@@ -161,28 +162,28 @@ class RunnableToRecipe(unittest.TestCase):
         self.tmpdir = tempfile.TemporaryDirectory(prefix=prefix)
 
     def test_runnable_to_recipe_noop(self):
-        runnable = nrunner.Runnable('noop', None)
+        runnable = Runnable('noop', None)
         recipe_path = os.path.join(self.tmpdir.name, 'recipe.json')
         runnable.write_json(recipe_path)
         self.assertTrue(os.path.exists(recipe_path))
-        loaded_runnable = nrunner.Runnable.from_recipe(recipe_path)
+        loaded_runnable = Runnable.from_recipe(recipe_path)
         self.assertEqual(loaded_runnable.kind, 'noop')
 
     def test_runnable_to_recipe_uri(self):
-        runnable = nrunner.Runnable('exec', '/bin/true')
+        runnable = Runnable('exec', '/bin/true')
         recipe_path = os.path.join(self.tmpdir.name, 'recipe.json')
         runnable.write_json(recipe_path)
         self.assertTrue(os.path.exists(recipe_path))
-        loaded_runnable = nrunner.Runnable.from_recipe(recipe_path)
+        loaded_runnable = Runnable.from_recipe(recipe_path)
         self.assertEqual(loaded_runnable.kind, 'exec')
         self.assertEqual(loaded_runnable.uri, '/bin/true')
 
     def test_runnable_to_recipe_args(self):
-        runnable = nrunner.Runnable('exec', '/bin/sleep', '0.01')
+        runnable = Runnable('exec', '/bin/sleep', '0.01')
         recipe_path = os.path.join(self.tmpdir.name, 'recipe.json')
         runnable.write_json(recipe_path)
         self.assertTrue(os.path.exists(recipe_path))
-        loaded_runnable = nrunner.Runnable.from_recipe(recipe_path)
+        loaded_runnable = Runnable.from_recipe(recipe_path)
         self.assertEqual(loaded_runnable.kind, 'exec')
         self.assertEqual(loaded_runnable.uri, '/bin/sleep')
         self.assertEqual(loaded_runnable.args, ('0.01', ))
@@ -194,7 +195,7 @@ class RunnableToRecipe(unittest.TestCase):
 class Runner(unittest.TestCase):
 
     def test_runner_noop(self):
-        runnable = nrunner.Runnable('noop', None)
+        runnable = Runnable('noop', None)
         runner_klass = runnable.pick_runner_class()
         runner = runner_klass()
         results = [status for status in runner.run(runnable)]
@@ -203,8 +204,8 @@ class Runner(unittest.TestCase):
         self.assertIn('time', last_result)
 
     def test_runner_exec(self):
-        runnable = nrunner.Runnable('exec-test', sys.executable,
-                                    '-c', 'import time; time.sleep(0.01)')
+        runnable = Runnable('exec-test', sys.executable,
+                            '-c', 'import time; time.sleep(0.01)')
         runner_klass = runnable.pick_runner_class()
         runner = runner_klass()
         results = [status for status in runner.run(runnable)]
@@ -220,8 +221,8 @@ class Runner(unittest.TestCase):
         self.assertIn('time', last_result)
 
     def test_runner_exec_test_ok(self):
-        runnable = nrunner.Runnable('exec-test', sys.executable,
-                                    '-c', 'import time; time.sleep(0.01)')
+        runnable = Runnable('exec-test', sys.executable,
+                            '-c', 'import time; time.sleep(0.01)')
         runner_klass = runnable.pick_runner_class()
         runner = runner_klass()
         results = [status for status in runner.run(runnable)]
@@ -239,7 +240,7 @@ class Runner(unittest.TestCase):
 
     @skipUnlessPathExists('/bin/false')
     def test_runner_exec_test_fail(self):
-        runnable = nrunner.Runnable('exec-test', '/bin/false')
+        runnable = Runnable('exec-test', '/bin/false')
         runner_klass = runnable.pick_runner_class()
         runner = runner_klass()
         results = [status for status in runner.run(runnable)]
@@ -256,7 +257,7 @@ class Runner(unittest.TestCase):
         self.assertIn('time', last_result)
 
     def test_runner_python_unittest_error(self):
-        runnable = nrunner.Runnable('python-unittest', 'error')
+        runnable = Runnable('python-unittest', 'error')
         runner_klass = runnable.pick_runner_class()
         runner = runner_klass()
         results = [status for status in runner.run(runnable)]
@@ -268,7 +269,7 @@ class Runner(unittest.TestCase):
         self.assertEqual(result['output'], output)
 
     def test_runner_python_unittest_empty_uri_error(self):
-        runnable = nrunner.Runnable('python-unittest', '')
+        runnable = Runnable('python-unittest', '')
         runner_klass = runnable.pick_runner_class()
         runner = runner_klass()
         results = [status for status in runner.run(runnable)]
@@ -298,7 +299,7 @@ echo 'not ok 2 - description 2'"""
         with open(tap_path, 'w', encoding='utf-8') as fp:
             fp.write(tap_script)
 
-        runnable = nrunner.Runnable('tap', '/bin/sh', tap_path)
+        runnable = Runnable('tap', '/bin/sh', tap_path)
         runner = runner_tap.TAPRunner()
         results = [status for status in runner.run(runnable)]
         last_result = results[-1]
@@ -318,7 +319,7 @@ echo 'ok 2 - description 2'"""
         with open(tap_path, 'w', encoding='utf-8') as fp:
             fp.write(tap_script)
 
-        runnable = nrunner.Runnable('tap', '/bin/sh', tap_path)
+        runnable = Runnable('tap', '/bin/sh', tap_path)
         runner = runner_tap.TAPRunner()
         results = [status for status in runner.run(runnable)]
         last_result = results[-1]
@@ -338,7 +339,7 @@ echo 'ok 2 - description 2'"""
         with open(tap_path, 'w', encoding='utf-8') as fp:
             fp.write(tap_script)
 
-        runnable = nrunner.Runnable('tap', '/bin/sh', tap_path)
+        runnable = Runnable('tap', '/bin/sh', tap_path)
         runner = runner_tap.TAPRunner()
         results = [status for status in runner.run(runnable)]
         last_result = results[-1]
@@ -358,7 +359,7 @@ echo 'ok 2 - description 2'"""
         with open(tap_path, 'w', encoding='utf-8') as fp:
             fp.write(tap_script)
 
-        runnable = nrunner.Runnable('tap', '/bin/sh', tap_path)
+        runnable = Runnable('tap', '/bin/sh', tap_path)
         runner = runner_tap.TAPRunner()
         results = [status for status in runner.run(runnable)]
         last_result = results[-1]
@@ -378,7 +379,7 @@ echo 'ok 2 - description 2'"""
         with open(tap_path, 'w', encoding='utf-8') as fp:
             fp.write(tap_script)
 
-        runnable = nrunner.Runnable('tap', '/bin/sh', tap_path)
+        runnable = Runnable('tap', '/bin/sh', tap_path)
         runner = runner_tap.TAPRunner()
         results = [status for status in runner.run(runnable)]
         last_result = results[-1]
@@ -394,8 +395,8 @@ echo 'ok 2 - description 2'"""
 class RunnerCommandSelection(unittest.TestCase):
 
     def setUp(self):
-        self.runnable = nrunner.Runnable('mykind',
-                                         'test_runner_command_selection')
+        self.runnable = Runnable('mykind',
+                                 'test_runner_command_selection')
 
     def test_is_task_kind_supported(self):
         cmd = ['sh', '-c',
@@ -417,8 +418,8 @@ class RunnerCommandSelection(unittest.TestCase):
 class PickRunner(unittest.TestCase):
 
     def setUp(self):
-        self.runnable = nrunner.Runnable('lets-image-a-kind',
-                                         'test_pick_runner_command')
+        self.runnable = Runnable('lets-image-a-kind',
+                                 'test_pick_runner_command')
 
     def test_pick_runner_command(self):
         runner = ['avocado-runner-lets-image-a-kind']
@@ -429,16 +430,16 @@ class PickRunner(unittest.TestCase):
         self.assertFalse(self.runnable.pick_runner_command({}))
 
 
-class Task(unittest.TestCase):
+class TaskTest(unittest.TestCase):
 
     def test_default_category(self):
-        runnable = nrunner.Runnable('noop', 'noop_uri')
-        task = nrunner.Task(runnable, 'task_id')
+        runnable = Runnable('noop', 'noop_uri')
+        task = Task(runnable, 'task_id')
         self.assertEqual(task.category, 'test')
 
     def test_set_category(self):
-        runnable = nrunner.Runnable('noop', 'noop_uri')
-        task = nrunner.Task(runnable, 'task_id', category='new_category')
+        runnable = Runnable('noop', 'noop_uri')
+        task = Task(runnable, 'task_id', category='new_category')
         self.assertEqual(task.category, 'new_category')
