+++ selftests/unit/test_runner_sysinfo.py
@@ -1,7 +1,7 @@
 import os
 import unittest
 
-from avocado.core.nrunner import Runnable
+from avocado.core.nrunner.runnable import Runnable
 from avocado.core.runners.sysinfo import SysinfoRunner
 from avocado.core.settings import settings
