+++ selftests/unit/test_tags.py
@@ -2,7 +2,7 @@ import stat
 import unittest
 
 from avocado.core import loader, tags
-from avocado.core.nrunner import Runnable
+from avocado.core.nrunner.runnable import Runnable
 from avocado.utils import script
 
 #: What is commonly known as "0664" or "u=rw,g=rw,o=r"
