+++ setup.py
@@ -311,11 +311,11 @@ if __name__ == '__main__':
           entry_points={
               'console_scripts': [
                   'avocado = avocado.core.main:main',
-                  'avocado-runner = avocado.core.nrunner:main',
-                  'avocado-runner-noop = avocado.core.nrunner:main',
-                  'avocado-runner-dry-run = avocado.core.nrunner:main',
-                  'avocado-runner-exec-test = avocado.core.nrunner:main',
-                  'avocado-runner-python-unittest = avocado.core.nrunner:main',
+                  'avocado-runner = avocado.core.nrunner.__main__:main',
+                  'avocado-runner-noop = avocado.core.runners.noop:main',
+                  'avocado-runner-dry-run = avocado.core.runners.dry_run:main',
+                  'avocado-runner-exec-test = avocado.core.runners.exec_test:main',
+                  'avocado-runner-python-unittest = avocado.core.runners.python_unittest:main',
                   'avocado-runner-avocado-instrumented = avocado.core.runners.avocado_instrumented:main',
                   'avocado-runner-tap = avocado.core.runners.tap:main',
                   'avocado-runner-asset = avocado.core.runners.requirement_asset:main',
@@ -402,10 +402,10 @@ if __name__ == '__main__':
                   ('avocado-instrumented = avocado.core.'
                    'runners.avocado_instrumented:AvocadoInstrumentedTestRunner'),
                   'tap = avocado.core.runners.tap:TAPRunner',
-                  'noop = avocado.core.nrunner:NoOpRunner',
-                  'dry-run = avocado.core.nrunner:DryRunRunner',
-                  'exec-test = avocado.core.nrunner:ExecTestRunner',
-                  'python-unittest = avocado.core.nrunner:PythonUnittestRunner',
+                  'noop = avocado.core.runners.noop:NoOpRunner',
+                  'dry-run = avocado.core.runners.dry_run:DryRunRunner',
+                  'exec-test = avocado.core.runners.exec_test:ExecTestRunner',
+                  'python-unittest = avocado.core.runners.python_unittest:PythonUnittestRunner',
                   'asset = avocado.core.runners.requirement_asset:RequirementAssetRunner',
                   'package = avocado.core.runners.requirement_package:RequirementPackageRunner',
                   'sysinfo = avocado.core.runners.sysinfo:SysinfoRunner',
